Packages: Extract dom package and expose it with Lerna
