+++ editor/components/multi-select-scroll-into-view/index.js
@@ -8,7 +8,7 @@ import scrollIntoView from 'dom-scroll-into-view';
  */
 import { Component } from '@wordpress/element';
 import { withSelect } from '@wordpress/data';
-import { getScrollContainer } from '@wordpress/utils';
+import { getScrollContainer } from '@wordpress/dom';
 
 /**
  * Internal dependencies
