+++ editor/components/observe-typing/index.js
@@ -8,7 +8,8 @@ import { includes } from 'lodash';
  */
 import { Component, compose } from '@wordpress/element';
 import { withSelect, withDispatch } from '@wordpress/data';
-import { isTextField, keycodes } from '@wordpress/utils';
+import { isTextField } from '@wordpress/dom';
+import { keycodes } from '@wordpress/utils';
 import { withSafeTimeout } from '@wordpress/components';
 
 const { UP, RIGHT, DOWN, LEFT, ENTER, BACKSPACE } = keycodes;
