+++ packages/dom/README.md
@@ -1,0 +1,13 @@
+# @wordpress/dom
+
+DOM utils module for WordPress.
+
+## Installation
+
+Install the module
+
+```bash
+npm install @wordpress/dom@next --save
+```
+
+<br/><br/><p align="center"><img src="https://s.w.org/style/images/codeispoetry.png?1" alt="Code is Poetry." /></p>
