+++ packages/dom/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "@wordpress/dom",
+	"version": "0.0.1",
+	"description": "DOM utils module for WordPress",
+	"author": "WordPress",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"dom",
+		"utils"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/dom/README.md",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/gutenberg.git"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"main": "src/index.js",
+	"dependencies": {
+		"element-closest": "2.0.2",
+		"lodash": "4.17.5",
+		"tinymce": "4.7.2"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
