--- utils/focus/index.js
@@ -1,4 +1,0 @@
-import * as focusable from './focusable';
-import * as tabbable from './tabbable';
-
-export { focusable, tabbable };
