+++ utils/index.js
@@ -1,14 +1,11 @@
-import * as focus from './focus';
 import * as keycodes from './keycodes';
 import * as viewPort from './viewport';
 import { decodeEntities } from './entities';
 
-export { focus };
 export { keycodes };
 export { decodeEntities };
 
 export * from './blob-cache';
-export * from './dom';
 export * from './mediaupload';
 export * from './terms';
 export * from './deprecation';
